# Implement neighborhood search based on CellListMap.jl