--- a/src/PointNeighbors.jl
+++ b/src/PointNeighbors.jl
@@ -4,7 +4,6 @@ using Reexport: @reexport
 
 using Adapt: Adapt
 using Atomix: Atomix
-using CellListMap: CellListMap
 using GPUArraysCore: AbstractGPUArray
 using KernelAbstractions: KernelAbstractions, @kernel, @index
 using LinearAlgebra: dot
@@ -18,11 +17,11 @@ include("nhs_trivial.jl")
 include("cell_lists/cell_lists.jl")
 include("nhs_grid.jl")
 include("nhs_precomputed.jl")
-include("celllistmap_nhs.jl")
 include("gpu.jl")
 
 export foreach_point_neighbor, foreach_neighbor
-export TrivialNeighborhoodSearch, GridNeighborhoodSearch, PrecomputedNeighborhoodSearch
+export TrivialNeighborhoodSearch, GridNeighborhoodSearch, PrecomputedNeighborhoodSearch,
+       CellListMapNeighborhoodSearch
 export DictionaryCellList, FullGridCellList
 export ParallelUpdate, SemiParallelUpdate, SerialUpdate
 export initialize!, update!, initialize_grid!, update_grid!