Based on #5.

I finally properly implemented the prototype in https://github.com/trixi-framework/TrixiParticles.jl/pull/174.